Modified main.css and birriaTacos.html files
I had added a background image but did not like it so I removed it, hence the main.css added to the commit. I had added it after placing it in the index and then removed. The birriaTacos.html file was modified because I added further list items to both the ordered and unordered lists. I also added <em> to the figcaption.